+++ optree/__init__.py
@@ -42,12 +42,14 @@ from optree.ops import (
     tree_transpose,
     tree_unflatten,
     treespec_children,
+    treespec_entries,
     treespec_is_leaf,
     treespec_is_prefix,
     treespec_is_strict_leaf,
     treespec_is_suffix,
     treespec_leaf,
     treespec_none,
+    treespec_paths,
     treespec_tuple,
 )
 from optree.registry import (
@@ -105,6 +107,8 @@ __all__ = [
     'prefix_errors',
     'treespec_is_prefix',
     'treespec_is_suffix',
+    'treespec_paths',
+    'treespec_entries',
     'treespec_children',
     'treespec_is_leaf',
     'treespec_is_strict_leaf',